refactor mixed/ll dicom converters
